Remove unnecessary AuthModule
