+++ modules/chat/index.ts
@@ -2,13 +2,13 @@ import { GraphQLModule } from '@graphql-modules/core';
 import { loadResolversFiles, loadSchemaFiles } from 'graphql-toolkit';
 import { UserModule } from '../user';
 import { ChatProvider } from './providers/chat.provider';
-import { AuthModule } from '../auth';
+import { CommonModule } from '../common';
 import { ProviderScope } from '@graphql-modules/di';
 
 export const ChatModule = new GraphQLModule({
   name: "Chat",
   imports: [
-    AuthModule,
+    CommonModule,
     UserModule,
   ],
   providers: [
