+++ modules/message/providers/message.provider.ts
@@ -6,7 +6,6 @@ import { Chat } from '../../../entity/Chat';
 import { ChatProvider } from '../../chat/providers/chat.provider';
 import { Message } from '../../../entity/Message';
 import { MessageType } from '../../../db';
-import { AuthProvider } from '../../auth/providers/auth.provider';
 import { UserProvider } from '../../user/providers/user.provider';
 
 @Injectable()
@@ -15,12 +14,11 @@ export class MessageProvider {
     private pubsub: PubSub,
     private connection: Connection,
     private chatProvider: ChatProvider,
-    private authProvider: AuthProvider,
     private userProvider: UserProvider,
   ) { }
 
   repository = this.connection.getRepository(Message);
-  currentUser = this.authProvider.currentUser;
+  currentUser = this.userProvider.currentUser;
 
   createQueryBuilder() {
     return this.connection.createQueryBuilder(Message, 'message');
